requirements.txt
-i https://pypi.org/simple
django==4.0.5
django-filter==23.5
django-cors-headers==3.13.0
django-extensions==3.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
python-decouple==3.6
pandas==2.1.4
psycopg2-binary==2.9.3
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyjwt==2.4.0; python_version >= '3.6'